+++ kubejs/client_scripts/remove_recipe_categories.js
@@ -1,0 +1,14 @@
+// RecipeViewerEvents.removeCategories((event) => {
+//     // 定义 removeRecipeCategories 数组
+//     const removeRecipeCategories = [
+//         "opolisutilities:drying_table",
+//         "opolisutilities:fluid_generator",
+//         "opolisutilities:resource_generator",
+//         "opolisutilities:rg2_blocks",
+//         "opolisutilities:rg2_speed_blocks",
+//         "opolisutilities:soaking_table",
+//     ];
+
+//     // removeRecipeCategories
+//     removeRecipeCategories.forEach((category) => event.remove(category));
+// });