+++ kubejs/server_scripts/remove_recipes.js
@@ -1,8 +1,6 @@
 // priority: -999
 
 ServerEvents.recipes((event) => {
-    console.log("#####", Ingredient.of("@opolisutilities").getItemIds());
-
     // 定义 removeRecipesUsedRecipeId 数组
     const removeRecipesUsedRecipeId = [
         //gtceu
@@ -11,6 +9,18 @@ ServerEvents.recipes((event) => {
         "gtceu:shaped/bowl",
         "gtceu:shaped/oak_planks_saw",
         "gtceu:shaped/oak_slab_saw",
+        "gtceu:shapeless/gravel_to_flint",
+        "gtceu:shaped/sugar",
+        "gtceu:shaped/quartz_sand",
+        "gtceu:shapeless/clay_block_to_dust",
+        "gtceu:shapeless/brick_block_to_dust",
+        "gtceu:shapeless/clay_ball_to_dust",
+        "gtceu:shapeless/brick_to_dust",
+        "gtceu:shaped/stone_slab_saw",
+        "gtceu:shaped/smooth_stone_slab_saw",
+        "gtceu:shaped/glass_pane",
+        "gtceu:shaped/mortar_grind_copper",
+        "gtceu:shapeless/bone_to_bone_meal",
 
         // opolisutilities
         "opolisutilities:sticks",
@@ -19,14 +29,72 @@ ServerEvents.recipes((event) => {
 
         // strainers
         "strainers:dirt",
+        "strainers:calcite_from_straining_diorite_in_strainers_purified_water_fluid",
 
         // farmersdelight
         "farmersdelight:flint_knife",
         "farmersdelight:iron_knife",
+
+        // farmersplus
+        "farmersplus:milling/cobbled_deepslate",
+        "farmersplus:milling/gravel",
+        "farmersplus:milling/cobblestone",
+        "farmersplus:milling/bone",
+
+        // minecraft
+        "minecraft:diorite",
+        "minecraft:bone_meal",
+
+        // botania
+        "botania:petal_white",
+        "botania:petal_white_double",
+        "botania:petal_orange",
+        "botania:petal_orange_double",
+        "botania:petal_magenta",
+        "botania:petal_magenta_double",
+        "botania:petal_light_blue",
+        "botania:petal_light_blue_double",
+        "botania:petal_yellow",
+        "botania:petal_yellow_double",
+        "botania:petal_lime",
+        "botania:petal_lime_double",
+        "botania:petal_pink",
+        "botania:petal_pink_double",
+        "botania:petal_gray",
+        "botania:petal_gray_double",
+        "botania:petal_light_gray",
+        "botania:petal_light_gray_double",
+        "botania:petal_cyan",
+        "botania:petal_cyan_double",
+        "botania:petal_purple",
+        "botania:petal_purple_double",
+        "botania:petal_blue",
+        "botania:petal_blue_double",
+        "botania:petal_brown",
+        "botania:petal_brown_double",
+        "botania:petal_green",
+        "botania:petal_green_double",
+        "botania:petal_red",
+        "botania:petal_red_double",
+        "botania:petal_black",
+        "botania:petal_black_double",
+
+        // create
+        "create:crafting/materials/copper_ingot",
+
+        // crossroads
+        "crossroads:smelting/dust_copper",
     ];
 
     // 定义 removeRecipesUsedItemInput 数组
-    const removeRecipesUsedItemInput = [];
+    const removeRecipesUsedItemInput = [
+        "#forge:ores",
+        "#forge:crushed_ores",
+        "#forge:purified_ores",
+        "#forge:pure_dusts",
+        "#forge:refined_ores",
+        "#forge:impure_dusts",
+    ];
 
     // 定义 removeRecipesUsedItemOutput 数组
     const removeRecipesUsedItemOutput = [
@@ -52,6 +120,13 @@ ServerEvents.recipes((event) => {
 
         // opolisutilities
         "opolisutilities:leafy_string",
+
+        // farmersplus
+        "farmersplus:millstone",
+
+        // compactcrafting
+        "compactcrafting:base",
+        "compactcrafting:projector_dish",
     ];
 
     // 定义 removeRecipesUsedItemInputAndOutput 数组
@@ -67,10 +142,21 @@ ServerEvents.recipes((event) => {
         "integrateddynamics:squeezer",
         "integrateddynamics:mechanical_drying_basin",
         "integrateddynamics:mechanical_squeezer",
+
+        // opolisutilities
+        "opolisutilities:drying_table",
+        "opolisutilities:fluid_generator",
+        "opolisutilities:resource_generator",
+        "opolisutilities:rg2_blocks",
+        "opolisutilities:rg2_speed_blocks",
+        "opolisutilities:soaking_table",
     ];
 
     // 定义 removeRecipesUsedTypeAndInput 数组
-    const removeRecipesUsedTypeAndInput = [];
+    const removeRecipesUsedTypeAndInput = [
+        { type: "minecraft:smelting", input: "#forge:raw_materials" },
+        { type: "minecraft:blasting", input: "#forge:raw_materials" },
+    ];
 
     // 定义 removeRecipesUsedTypeAndOutput 数组
     const removeRecipesUsedTypeAndOutput = [];
@@ -102,13 +188,6 @@ ServerEvents.recipes((event) => {
         // remove output
         event.remove({ output: item });
     });
-
-    // // c:hidden_from_recipe_viewers
-    // // remove input
-    // event.remove({ input: Ingredient.of("#c:hidden_from_recipe_viewers") });
-
-    // // remove output
-    // event.remove({ output: Ingredient.of("#c:hidden_from_recipe_viewers") });
 
     // removeRecipesUsedType
     removeRecipesUsedType.forEach((type) => {